% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{delete_transcription_job}
\alias{delete_transcription_job}
\title{Deletes a previously submitted transcription job as well as any other generated results such as the transcription, models, and so on}
\usage{
delete_transcription_job(TranscriptionJobName)
}
\arguments{
\item{TranscriptionJobName}{The name of the transcription job to be deleted.}
}
\description{
Deletes a previously submitted transcription job as well as any other generated results such as the transcription, models, and so on.
}
\examples{

}
