% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{delete_vocabulary}
\alias{delete_vocabulary}
\title{Deletes a vocabulary from Amazon Transcribe}
\usage{
delete_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{The name of the vocabulary to delete.}
}
\description{
Deletes a vocabulary from Amazon Transcribe.
}
\examples{

}
