% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Gets information about a vocabulary}
\usage{
get_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{The name of the vocabulary to return information about. The name is case-sensitive.}
}
\description{
Gets information about a vocabulary. Note that vocabularies for en-AU, en-UK, and fr-CA languages that are in preview are not available. In the console, the vocabulary section will be greyed-out and SDK will return error message.
}
\examples{

}
