% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{list_transcription_jobs}
\alias{list_transcription_jobs}
\title{Lists transcription jobs with the specified status}
\usage{
list_transcription_jobs(Status = NULL, JobNameContains = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{Status}{When specified, returns only transcription jobs with the specified status.}

\item{JobNameContains}{When specified, the jobs returned in the list are limited to jobs whose name contains the specified string.}

\item{NextToken}{If the result of the previous request to \code{ListTranscriptionJobs} was truncated, include the \code{NextToken} to fetch the next set of jobs.}

\item{MaxResults}{The maximum number of jobs to return in the response. If there are fewer results in the list, this response contains only the actual results.}
}
\description{
Lists transcription jobs with the specified status.
}
\section{Accepted Parameters}{
\preformatted{list_transcription_jobs(
  Status = "IN_PROGRESS"|"FAILED"|"COMPLETED",
  JobNameContains = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

