% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_package.R
\docType{package}
\name{paws.transcribeservice-package}
\alias{paws.transcribeservice}
\alias{paws.transcribeservice-package}
\title{paws.transcribeservice: Amazon Transcribe Service}
\description{
Operations and objects for transcribing speech to
text.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
