% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{start_transcription_job}
\alias{start_transcription_job}
\title{Starts an asynchronous job to transcribe speech to text}
\usage{
start_transcription_job(TranscriptionJobName, LanguageCode,
  MediaSampleRateHertz = NULL, MediaFormat, Media,
  OutputBucketName = NULL, Settings = NULL)
}
\arguments{
\item{TranscriptionJobName}{[required] The name of the job. You can't use the strings "." or ".." in the job name. The name must be unique within an AWS account.}

\item{LanguageCode}{[required] The language code for the language used in the input media file.}

\item{MediaSampleRateHertz}{The sample rate, in Hertz, of the audio track in the input media file.}

\item{MediaFormat}{[required] The format of the input media file.}

\item{Media}{[required] An object that describes the input media for a transcription job.}

\item{OutputBucketName}{The location where the transcription is stored.

If you set the \code{OutputBucketName}, Amazon Transcribe puts the transcription in the specified S3 bucket. When you call the GetTranscriptionJob operation, the operation returns this location in the \code{TranscriptFileUri} field. The S3 bucket must have permissions that allow Amazon Transcribe to put files in the bucket. For more information, see \href{https://docs.aws.amazon.com/transcribe/latest/dg/access-control-managing-permissions.html#auth-role-iam-user}{Permissions Required for IAM User Roles}.

If you don't set the \code{OutputBucketName}, Amazon Transcribe generates a pre-signed URL, a shareable URL that provides secure access to your transcription, and returns it in the \code{TranscriptFileUri} field. Use this URL to download the transcription.}

\item{Settings}{A \code{Settings} object that provides optional settings for a transcription job.}
}
\description{
Starts an asynchronous job to transcribe speech to text. Note that en-AU, en-UK, and fr-CA languages are in preview and are only available to whitelisted customers.
}
\section{Accepted Parameters}{
\preformatted{start_transcription_job(
  TranscriptionJobName = "string",
  LanguageCode = "en-US"|"es-US"|"en-AU"|"fr-CA"|"en-UK",
  MediaSampleRateHertz = 123,
  MediaFormat = "mp3"|"mp4"|"wav"|"flac",
  Media = list(
    MediaFileUri = "string"
  ),
  OutputBucketName = "string",
  Settings = list(
    VocabularyName = "string",
    ShowSpeakerLabels = TRUE|FALSE,
    MaxSpeakerLabels = 123,
    ChannelIdentification = TRUE|FALSE
  )
)
}
}

