% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{update_vocabulary}
\alias{update_vocabulary}
\title{Updates an existing vocabulary with new values}
\usage{
update_vocabulary(VocabularyName, LanguageCode, Phrases)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary to update. The name is case-sensitive.}

\item{LanguageCode}{[required] The language code of the vocabulary entries.}

\item{Phrases}{[required] An array of strings containing the vocabulary entries.}
}
\description{
Updates an existing vocabulary with new values. The \code{UpdateVocabulary} operation overwrites all of the existing information with the values that you provide in the request. Note that vocabularies for en-AU, en-UK, and fr-CA languages that are in preview are not available. In the console, the vocabulary section will be greyed-out and SDK will return error message.
}
\section{Accepted Parameters}{
\preformatted{update_vocabulary(
  VocabularyName = "string",
  LanguageCode = "en-US"|"es-US"|"en-AU"|"fr-CA"|"en-UK",
  Phrases = list(
    "string"
  )
)
}
}

