% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{create_vocabulary}
\alias{create_vocabulary}
\title{Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file}
\usage{
create_vocabulary(VocabularyName, LanguageCode, Phrases)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary. The name must be unique within an AWS account. The name is case-sensitive.}

\item{LanguageCode}{[required] The language code of the vocabulary entries.}

\item{Phrases}{[required] An array of strings that contains the vocabulary entries.}
}
\description{
Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file. Note that vocabularies for en-AU, en-UK, and fr-CA languages that are in preview are not available. In the console, the vocabulary section will be greyed-out and SDK will return error message.
}
\section{Accepted Parameters}{
\preformatted{create_vocabulary(
  VocabularyName = "string",
  LanguageCode = "en-US"|"es-US"|"en-AU"|"fr-CA"|"en-UK",
  Phrases = list(
    "string"
  )
)
}
}

