% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.transcribeservice_operations.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Gets information about a vocabulary}
\usage{
get_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the vocabulary to return information about. The name is case-sensitive.}
}
\description{
Gets information about a vocabulary.
}
\section{Accepted Parameters}{
\preformatted{get_vocabulary(
  VocabularyName = "string"
)
}
}

